// src/store/configs/types.ts
enum types {
	TOGGLE_FOLLOWED = "TOGGLE_FOLLOWED",
	SORT_FOLLOWED = "SORT_FOLLOWED",
	SET_WIDTH = "SET_WIDTH",
	RESET_WIDTH = "RESET_WIDTH",
	SET_NOTIFICATION = "SET_NOTIFICATION",
	SET_LOCALE = "SET_LOCALE",
	TOGGLE_SHOW_BOSS_IMAGE = "TOGGLE_SHOW_BOSS_IMAGE",
	TOGGLE_SHOW_USER_IMAGE = "TOGGLE_SHOW_USER_IMAGE",
	SET_TIME_FORMATION = "SET_TIME_FORMATION",
	SET_LIST_MODE = "SET_LIST_MODE"
}

export default types